compose image service and upload service in attachment service
