+++ src/app/admin/season/season.html
@@ -22,7 +22,6 @@
     </mhs-attachment>
   </div>
 
-
   <table ng-if=" $ctrl.seasonTeams.length != 0" class="table table-sm table-striped table-hover">
     <caption class="table-header">{{'TEAMS_IN_SEASON' | translate}}</caption>
     <thead>
