+++ src/app/hybrid-app.module.ts
@@ -49,7 +49,6 @@ import { FacebookShareComponent } from './facebook-share/facebook-share.componen
 import { GameProgressComponent } from './admin/game-progress/game-progress.component';
 import { RoundPanelComponent } from './admin/game-progress/round-panel/round-panel.component';
 import { BackupDirective } from './shared/backup.directive';
-import { ImageService } from './services/image-service/image.service';
 
 import { DatePickerComponent } from './admin/date-picker/date-picker.component';
 import { TimePickerComponent } from './admin/time-picker/time-picker.component';
@@ -67,7 +66,7 @@ import { ConvertService } from './services/convert-service/convert.service.upgra
 import { GameBuildService } from './services/game-build-service/game-build.service.upgrade';
 import { RoundStatusService } from './services/round-service/round-status.service.upgrade';
 import { AttachmentComponent } from './admin/attachment/attachment.component';
-import { UploadService } from "./services/upload-service/upload.service";
+import { AttachmentService } from './services/attachment-service/attachment.service';
 
 defineLocale('ru', ru);
 defineLocale('en', enGb);
@@ -169,8 +168,7 @@ export function HttpLoaderFactory(http: HttpClient) {
     ConvertService,
     GameBuildService,
     RoundStatusService,
-    ImageService,
-    UploadService
+    AttachmentService,
   ],
 })
 export class HybridAppModule {
